backend/requirements.txt
fastapi
uvicorn
motor
beanie
python-dotenv
pydantic
pydantic-settings
email-validator
openai
bcrypt
pyjwt
python-jose
